requirements.txt
Flask
requests
patool
